requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
Naked==0.1.31
protobuf==3.20.1
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.14.1
PyMySQL==1.0.2
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
flask_cors==3.0.6
